feat(contests): added separate app for contests, moved Yandex Contest API to the app
