+++ lms/settings/extended.py
@@ -17,6 +17,7 @@ INSTALLED_APPS += [
     'surveys.apps.SurveysConfig',
     'info_blocks.apps.InfoBlocksConfig',
     'code_reviews.apps.CodeReviewsConfig',
+    'contests.apps.ContestsConfig',
     'faq.apps.FAQConfig',
     'ckeditor',
     'ckeditor_uploader',
